test(vscode): move unit tests into own test suite (#17533)
> This PR reorganizes the VSCode extension tests by separating unit tests into their own test suite, distinct from integration tests. This allows unit tests to run independently and faster without requiring LSP server initialization.
